+++ examples/pathogen_evolution/pathogen_evolution.jl
@@ -29,6 +29,8 @@ function run(seed::Int64, t_vec::Vector{Float64})
         receiveContactCoefficient=s::String -> 0.0,
     )
 
+    hos_type = jOpqua.newHostType("hos_type")
+
     pop_type = jOpqua.newPopulationType(
         "pop_type",
         clearance_coefficient=1.0,
@@ -44,7 +46,7 @@ function run(seed::Int64, t_vec::Vector{Float64})
     # Setup
     model = jOpqua.newModel()
     pop = jOpqua.newPopulation!("pop", pop_type, model)
-    jOpqua.addHostsToPopulation!(num_hosts, host_genome, pop, model)
+    jOpqua.addHostsToPopulation!(num_hosts, host_genome, hos_type, pop, model)
     pat = jOpqua.newPathogen!(start_genome, pop, pat_type)
 
     for h in 1:num_infected