+++ changelog.md
@@ -5,7 +5,6 @@ KNOWN ISSUES:
 TODO:
 - Make `Host` sequence functions that affect event coefficients, I think those
 are missing?
-- Create `HostType` struct to house `Host`-specific parameters
 - Organize non-sampling coefficients and their functions into matrices and
 vectors
 - Rename infection probability to transmission efficiency
@@ -32,6 +31,7 @@ TODO: Not debugged:
 
 ## 3 March 2025
 - Added nonsampling coefficient vector to `Host` and some machinery to update it
+- Create `HostType` struct to house `Host`-specific parameters
 
 ## 2 March 2025
 - Corrected a couple references from to `NUM_COEFFICIENTS` for